applying patches from tracker (5 Issues working with files bigger than 4Gb - ID: 3312690) - patch 02-IntOverflow-Get_NCells.patch

+++ saga-gis/src/modules/grid/grid_gridding/Interpolation_NaturalNeighbour.cpp
@@ -135,9 +135,9 @@ bool CInterpolation_NaturalNeighbour::Interpolate(void)
 	delaunay	*pTIN	= delaunay_build(n, pSrc, 0, NULL, 0, NULL);
 
 	//-----------------------------------------------------
-	double	*xDst	= (double *)SG_Malloc(m_pGrid->Get_NCells() * sizeof(double));
-	double	*yDst	= (double *)SG_Malloc(m_pGrid->Get_NCells() * sizeof(double));
-	double	*zDst	= (double *)SG_Malloc(m_pGrid->Get_NCells() * sizeof(double));
+	double	*xDst	= (double *)SG_Malloc((long) m_pGrid->Get_NCells() * sizeof(double));
+	double	*yDst	= (double *)SG_Malloc((long) m_pGrid->Get_NCells() * sizeof(double));
+	double	*zDst	= (double *)SG_Malloc((long) m_pGrid->Get_NCells() * sizeof(double));
 
 	for(y=0, i=0, p.y=m_pGrid->Get_YMin(); y<m_pGrid->Get_NY() && Set_Progress(y, m_pGrid->Get_NY()); y++, p.y+=m_pGrid->Get_Cellsize())
 	{
