+++ saga-gis/src/modules_terrain_analysis/terrain_analysis/ta_morphometry/SurfaceSpecificPoints.cpp
@@ -396,7 +396,7 @@ void CSurfaceSpecificPoints::Do_FlowDirection2(CSG_Grid *pGrid, CSG_Grid *pResul
 
 	Do_FlowDirection(&Grid, pResult);
 
-	for(int n=0; n<Get_NCells(); n++)
+	for(long n=0; n<Get_NCells(); n++)
 	{
 		pResult->Add_Value(n, -Result.asInt(n));
 	}
